🏗️ Refactor: 스켈레톤 폴더 생성
