+++ src/components/skeletons/SkeletonCard.jsx
