+++ src/components/skeletons/SkeletonDetail.jsx
